move theme under styles
